+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 2.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Sort gate slots in get_free_slots method to better handle also pauses.
+  [cekk]
 
 
 2.1.0 (2023-10-11)